Much progress toward implementing new irrigation module.
It may be that the new irrigation module is incompatible with the Thornthwaite-Mather
soil moisture balance, in particular, the SM retention tables. With two-stage evap + transp,
evap is limited to the top layers of soil while transpiration can continue based on the needs
of the plant. This may require a new soil-moisture module.