Geographic transformations enabled for the four major input grids.
PROJ4 was modified and linked to swb, allowing for coordinate transformations to be performed on the fly. This allows ASCII grids of disparate projection and resolution to be included in a swb run without any further manipulation by external GIS packages. Also, a second form of the grid_Create function was added, which eliminates the need to define the upper right-hand project coordinates; instead, the upper right-hand coordinates are calculated on the basis of the user-provided grid cell resolution.