Revisit irrigation module; add required *.csv files. 'make check' passes for both test cases.
It appears that the irrigation test was failing owing to the treatment of excess water related to inefficiencies in application. This excess water is assumed to make it into the root zone; the CROPWAT program does *not* make this assumption: excess water generated in CROPWAT is assumed to bypass the root zone.