BUGFIX: dynamic landuse option fixed.
Dynamic landuse option was not working because the lMissingFilesAreAllowed flag was never set properly. This flag is now set to TRUE following the first successful read of a landuse file. This has the effect of allowing a different landuse file to be read in on any following day of the simulation.