Add logic to scan for incorrect number of data columns in the landuse lookup table.
If the HORTON interception parameters are added to the landuse lookup table but the INTERCEPTION_METHOD HORTON directive is not added to the control file, SWB will read in the "b" and "n" interception parameters as though they were rooting depth parameters. This logic will check to see whether the landuse file in use has "extra" columns of data present once all of the assumed landuse table values have been read in.